pyproject.toml
[tool.poetry]
name = "kbsb_api"
version = "1.0.0"
description = "API Website FRBE-KBSB-KSB"
authors = ["Ruben Decrop <ruben.decrop@frbe-kbsb-ksb.be>"]
license = "GPLv2"

[tool.poetry.dependencies]
python = "^3.10"
aiofiles = "22.1"
bcrypt = "^3.2.2"
colorlog = "^6.6.0"
dnspython = "^2.2.1"
fastapi = "0.87"
google = "^3.0.0"
google-api-python-client = "^2.73.0"
google-auth = "^2.16.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.1"
google-cloud-secret-manager = "^2.15.0"
google-cloud-storage = "^2.7.0"
gunicorn = "^20.1.0"
markdown2 = "^2.4.3"
motor = "^3.0.0"
passlib = "^1.7.4"
pymongo = "^4.1.1"
python-dateutil = "^2.8.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
PyYAML = "^6.0"
requests = "^2.27.1"
uvicorn = "0.19"
SQLAlchemy = "^1.4.36"
PyMySQL = "^1.0.2"
python-slugify = "^6.1.2"
Jinja2 = "^3.1.2"
weasyprint = "^57.2"
pandas = "^1.4.3"
openpyxl = "^3.0.10"
reddevil = {path = "libs/reddevil-3.0.4-py3-none-any.whl"}
google-api-core = "^2.11.0"
pathspec = "^0.10.3"

[tool.poetry.dev-dependencies]
asyncio = "^3.4.3"
black = "^22.3.0"
httpx = "^0.22.0"
mypy = "^0.950"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"

[tool.poetry.group.dev.dependencies]
toml = "^0.10.2"
click = "^8.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
